refactor proposal memo report package router
